Move gctf and ictf gamemodes to own folder
The idea is that every gamemode that can be set with sv_gametype
has a own folder
