+++ src/game/server/gamecontext.cpp
@@ -34,10 +34,10 @@
 #include "gamemodes/instagib/bolofng.h"
 #include "gamemodes/instagib/boomfng.h"
 #include "gamemodes/instagib/fng.h"
-#include "gamemodes/instagib/gctf.h"
+#include "gamemodes/instagib/gctf/gctf.h"
 #include "gamemodes/instagib/gdm.h"
 #include "gamemodes/instagib/gtdm.h"
-#include "gamemodes/instagib/ictf.h"
+#include "gamemodes/instagib/ictf/ictf.h"
 #include "gamemodes/instagib/idm.h"
 #include "gamemodes/instagib/itdm.h"
 #include "gamemodes/instagib/solofng.h"
