+++ src/game/server/gamemodes/instagib/gctf/gctf.h
@@ -1,5 +1,5 @@
-#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_GCTF_H
-#define GAME_SERVER_GAMEMODES_INSTAGIB_GCTF_H
+#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_GCTF_GCTF_H
+#define GAME_SERVER_GAMEMODES_INSTAGIB_GCTF_GCTF_H
 
 #include <game/server/gamemodes/instagib/ctf.h>
