+++ src/game/server/gamemodes/instagib/ictf/ictf.cpp
