+++ src/game/server/gamemodes/instagib/ictf/ictf.h
@@ -1,5 +1,5 @@
-#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_ICTF_H
-#define GAME_SERVER_GAMEMODES_INSTAGIB_ICTF_H
+#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_ICTF_ICTF_H
+#define GAME_SERVER_GAMEMODES_INSTAGIB_ICTF_ICTF_H
 
 #include <game/server/gamemodes/instagib/ctf.h>
